cleanup and bug fixes
fixed a bug where ifndef pin of led strip data did not include the settings file for the actual pinout. making the led strip not light up.

added a funtion that stops transmitting pixels when state is stagnant.

updated mapi works now.

added brightness knob updates brightness when knob2 is pressed,

+++ SK6812rgbw driver/SK6812rgbw driver/timer.h
@@ -10,6 +10,6 @@
 #define TIMER_H_
 
 void init_timer0(void);
-uint16_t Get_time_ms(void);
+uint32_t Get_time_ms(void);
 
 #endif /* TIMER_H_ */
